userIds saved with posts + only post author can edit/delete

+++ backend/routes/posts.js
@@ -10,7 +10,8 @@ router.post('', checkAuth, (req, res, next) => {
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        image: req.body.image
+        image: req.body.image,
+        creator: req.userData.userId
     })
     post.save()
     .then(savedPost => {
@@ -22,9 +23,15 @@ router.post('', checkAuth, (req, res, next) => {
 })
 //delete saved posts
 router.delete('/:id', checkAuth, (req, res, next) => {
-    Post.deleteOne({_id:  req.params.id})
-    .then(result => console.log('deleted'))
-    res.status(200).json({message: 'Post deleted'})
+    Post.deleteOne({_id:  req.params.id, creator: req.userData.userId})
+    .then(result => {
+        console.log(result)
+        if(result.n > 0) {
+            res.status(200).json({message: 'Post deleted'})
+        } else {
+            res.status(401).json({message: 'Not authorized!'})
+        }
+    })
 })
 // put edited post
 router.put('/:id', checkAuth, (req, res, next) => {
@@ -34,10 +41,13 @@ router.put('/:id', checkAuth, (req, res, next) => {
         image: req.body.image,
         _id: req.body.id
     })
-    Post.updateOne({_id: post._id}, post)
+    Post.updateOne({_id: post._id, creator: req.userData.userId}, post)
     .then(result => {
-        console.log('updated post')
-        res.status(200).json({message: 'Post update successful'})
+        if(result.nModified > 0) {
+            res.status(200).json({message: 'Post update successful'})
+        } else {
+            res.status(401).json({message: 'Not authorized!'})
+        }
     })
     .catch(error => {
         console.log(error, 'in appp.put')
